requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-groq==0.3.1
langchain-text-splitters==0.3.7
langsmith==0.3.19
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.0.2
orjson==3.10.16
packaging==24.2
propcache==0.3.1
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.0.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0